Renames 'utils' to 'components'
