+++ src/components/menus.js
@@ -94,9 +94,7 @@ module.exports = {
       label: 'Check for Updates',
       click: function() {
         updater.check(manifest, function(error, newVersionExists, newManifest) {
-          if (error) {
-            console.error(error);
-          } else if (newVersionExists) {
+          if (error || newVersionExists) {
             updater.prompt(win, error, newVersionExists, newManifest);
           } else {
             win.window.alert('You\'re using the latest version: ' + manifest.version);
