+++ src/components/platform.js
